Refactor server out of index file
